feat: remove previous components and build basic AuthenicationSlice
