+++ README.md
@@ -4,19 +4,19 @@ My note system
 
 ## Technology stack
 
+- [x] React
 - [x] Next.js (deploy on Vercel)
 - [x] TypeScript
-- [x] React
 - [x] ESLint
-- [x] Git hook (Husky)
-- [x] Firebase for database and authentication
 - [x] Redux Toolkit, Redux thunk
+- [x] Git hook (Husky)
+- [x] MaterialUI
+- [x] Firebase for database and authentication
 - [x] Unit Testing with Jest / React Testing Library / Testing Library jest-dom / Redux mock store
 - [x] End-to-End Testing with TestCafe 
-- [x] MaterialUI
 
 ## Todo
 
-- [ ] Navigator
+- [ ] Authentication
 - [ ] Main Page and Navigator
 - [ ] Workout Page
