--- public/index_template.html
@@ -1,12 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <link rel="icon" type="image/png" href="<%=require('./icon.png')%>" />
-    <title><%=htmlWebpackPlugin.options.title%></title>
-  </head>
-  <body>
-    <main id="root"></main>
-  </body>
-</html>
