--- src/app/store.ts
@@ -1,10 +1,0 @@
-import { configureStore } from "@reduxjs/toolkit"
-import counterReducer from "game-note/features/example/couter/counterSlice"
-
-const store = configureStore({
-  reducer: { counter: counterReducer }
-})
-
-export default store
-export type RootState = ReturnType<typeof store.getState>
-export type AppDispatch = typeof store.dispatch
