--- src/features/authentication/Login.style.css
@@ -1,113 +1,0 @@
-.loginComponent {
-  background-color: var(--block-background-color);
-  margin: 0 auto;
-  width: 20rem;
-  padding: 1rem 1rem 0.8rem 1rem;
-  border-radius: 8px;
-  border-color: var(--secondary-color);
-  border-style: solid;
-}
-
-.loginComponent form {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-}
-
-.inputBlock {
-  display: flex;
-  flex-direction: column;
-  align-items: flex-start;
-  margin-bottom: 0.4rem;
-  width: 100%;
-}
-
-.inputBlock input {
-  background-color: var(--input-background-color);
-  border-top-left-radius: 3px;
-  border-top-right-radius: 3px;
-  border-top: none;
-  border-left: none;
-  border-right: none;
-  border-bottom-style: dashed;
-  padding: 0.5rem;
-  font-size: 1.1em;
-  width: inherit;
-  box-sizing: border-box;
-  transition: all 0.2s ease;
-}
-
-.inputBlock input:focus {
-  border: none;
-  box-shadow: 1px 4px 8px 2px black;
-  outline: solid 2px var(--outline-color);
-  transition: all 0.2s ease;
-}
-
-.inputBlock label {
-  margin-bottom: 0.1rem;
-}
-
-.passwordInputBlock {
-  display: flex;
-  width: 100%;
-  margin-bottom: 0.1rem;
-  position: relative;
-  align-items: center;
-}
-
-.loginButton {
-  align-self: center;
-  margin-top: 0.3rem;
-  padding: 0.2rem;
-}
-
-.toggleVisibleButton {
-  cursor: pointer;
-  position: absolute;
-  right: 0;
-  height: 0.9em;
-  margin-right: 0.2em;
-  z-index: 1;
-}
-
-.errorMessageBlock {
-  margin-top: 0.5em;
-  color: red;
-  border-width: 1px;
-  border-style: solid;
-  border-color: black;
-  padding: 4px;
-  border-radius: 2px;
-}
-
-/* mobile style */
-@media (max-width: 465px) {
-  h1 {
-    padding: 10vh 0 3vh 0;
-  }
-
-  .loginComponent {
-    max-width: 100vw;
-    width: auto;
-    box-sizing: content-box;
-    margin: 0;
-  }
-
-  .loginComponent form {
-    align-items: flex-start;
-    box-sizing: content-box;
-  }
-
-  .loginComponent input {
-    width: inherit;
-    box-sizing: border-box;
-    font-size: 1.5em;
-    padding-top: 0.2rem;
-    padding-bottom: 0.2rem;
-  }
-
-  .toggleVisibleButton {
-    height: 1.1em;
-  }
-}
