+++ src/features/authentication/Login.tsx
@@ -1,0 +1,3 @@
+export function Login () {
+  return <p>Login Container</p>
+}
