--- src/shared/hooks/useRenderLog.js
@@ -1,11 +1,0 @@
-/* eslint-disable no-console */
-import React from "react"
-
-export function useRenderLog (componentName) {
-  React.useEffect(() => {
-    console.log(`${componentName} render`)
-    return () => {
-      console.log(`${componentName} render cleanup`)
-    }
-  })
-}
