+++ src/shared/store.ts
@@ -1,0 +1,15 @@
+import { TypedUseSelectorHook, useDispatch, useSelector } from "react-redux"
+import { configureStore } from "@reduxjs/toolkit"
+import authenticationSlice from "game-note/features/authentication/authenticationSlice"
+
+const store = configureStore({
+  reducer: { authentication: authenticationSlice }
+})
+
+export default store
+
+export type RootState = ReturnType<typeof store.getState>
+export type AppDispatch = typeof store.dispatch
+
+export const useAppDispatch = () => useDispatch<AppDispatch>()
+export const useAppSelector: TypedUseSelectorHook<RootState> = useSelector
